Move search code back to activity
